feat(qa): support for collecting structured logs
The general development direction here is to have support for:

1. generating golden files containing structured logs using
`./cmd/qatool` (to be implemented soon);

2. running the QA suite as part of `go test ./...`, collecting
the structured logs, and comparing with golden files (also
accounting for some variability, i.e., the comparison won't
be a dumb comparison for equality).

To this end, the first step is to create the conditions for
collecting the structured logs generated by running RBMK without
causing data races and in a way such that we can reuse the
same code for both purposes 1 and 2 above.

This diff implements that, by:

1. allowing to override the stdout used for structured logging;

2. passing `+logs` to all tests to actually emit logs;

3. extending `scenario.Run` to return the logs to the caller.