fix(dig curl): enforce max operation runtime (#9)
Use five seconds for dig without making it configurable. For now, this
default should be good enough. For DNS-over-HTTPS, ensure that the
timeout also applies to reading the body by setting client.Timeout.

For curl, implement the `--max-time` flag, default it to 30s, and ensure
it also applied to reading the body using client.Timeout.

Also, ensure we update the documentation accordingly.