refactor: update to rbmk-project/common/clitools@v0.10.0
The new clitools uses dependency injection to provide the command
Environment for Command.Help and Command.Main.

In turn, the Environment provides the stdout and the stderr.

Update our CLI commands and take advantage of this to simplify how
we make the stdout testable in the `./internal/qa` package.

See https://github.com/rbmk-project/common/pull/13