feat: implement the `rbmk stun` command (#22)
This command is a fundamental building block for research, because it
allows us to know our external IP address(es) and port(s). By analysing
this information, we're able to know the ASN, country code, and possibly
region from which we're running measurements. As noted in the new
`README.txt`, for accurate measurements, we would need to send binding
requests using both IPv4 and IPv6 STUN targets explicitly. Otherwise, if
using domain names, we would only know either the IPv4 or the IPv6
public address, depending on which STUN server address is used.