feat: optionally render docs using markdown
Documentation is more accessible when rendering markdown to the
terminal, however, the binary is larger.

So, start making this possible but, at the same time, allow building
without markdown rendering support for smaller binaries.

This diff only modifies the tutorial subcommand to this effect, but
we are going to modify also the rest of the commands soon.