fix: avoid garbled shell output with background commands
Before this diff was commited, the following script:

```bash
(./rbmk dig +short +udp=wait-duplicates example.com | ./rbmk pipe write addrs) &

./rbmk pipe read --writers 1 addrs | ./rbmk ipuniq

wait
```

was causing garbled output including ANSI escape sequences emitted by the
`rbmk pipe` command. We traced the problem to automatic detection of the
terminal capabilities by the markdown rendering library.

In turn, the issue was that we were always rendering the markdown for
cliutils.CommandWithSubCommands even when we were not needing it.

To fix this issue, we:

1. adapted rbmk-project/common (in https://github.com/rbmk-project/common/pull/15)

2. updated `./internal/markdown` and commands using CommandWithSubCommands
to use conditiional markdown rendering

After these changes, the issue has disappeared.