refactor: use common/fsx filesystem abstraction (#43)
This diff refactors the codebase to use the common/fsx filesystem
abstraction whenever it's trivial to do so.

Using this abstraction opens up the possibility of writing unit tests
for this package, rather than relying on the filesystem.

In principle, this change also opens up the possibility of running rbmk
commands as internal commands inside the built-in shell.

Speaking of which, I did not remove the shell's dependency on the `os`
package since doing that also requires to mutate the environment
variables, which probably we don't need if we transform all commands
into internal shell commands. So, regarding that, we will basically see
what to do at a later time rather than now.

While there, upgrade dependencies.