feat: add example showing a stun lookup
The general idea about this example is that of showing how we can
perform a STUN lookup. However, there are these additional reasons
for creating this example right now:

1. I want to ensure it is possible to use RBMK to perform relatively
complex tasks by writing specific shell scripts (that is, are we
missing any feature that would be useful?)

2. I want to figure out how to extend the tutorial, and the best way
to do this is to use the tool myself and experiment.