feat(nc): implement the `-T, --option` flag
The short flag is compatible with OpenBSD nc(1) and allows to
specify the `noverify` option to skip TLS verification.

In turn, skipping TLS verification is needed to avoid errors or
warnings (depending on `--measure`) when using `rbmk nc` to check
to SNI based blocking using arbitrary remote TLS servers as the
test servers (e.g., using `www.example.com:443` to check whether
a rule in the network blocks the `www.instagram.com` SNI).

Strictly speaking, we don't need this flag since the test is
okay also with verification enabled. Additionally, one can say
that expecting a verification error is actually the expected
state and disabling verification could miss a MITM. That said,
consider also that a test such as SNI blocking depends on the
assumptions you make. If you assume that MITM is not a concern,
then it leads to nicer UI to avoid emitting any error. Also,
do not miss the fact that TLS MITM could still be detected when
post-processing the measurement results.

While there, add missing documentation comments to intro.go.