fix(stun): the flag is --max-time not --maxtime (#54)
This commit fixes a defect in the `rbmk stun` command where the
`--max-time DURATION` flag was incorrectly parsed as `--maxtime` by the
command line flags parsing code. I don't consider this change a breaking
change given that the flag was clearly and correctly documented as
`--max-time` by `rbmk stun --help`.