chore: merge x/netcore into the rbmk repository
The overall purpose for this diff is to simplify the development
environment and reduce chores for me.

My original setup was OK for a professionally maintained project.
However, it is overkill for a hobby project.

This diff is based on https://github.com/rbmk-project/x/commit/5718c136805cdf80d8b2adbc973353831ce45433.