chore: merge common into the rbmk repository
This diff is based on https://github.com/rbmk-project/common/commit/443e41e5d38260ff6620be93e98184f41be40402.

The overall purpose for this diff is to simplify the development
environment and reduce chores for me. My original setup was OK for
a professionally maintained project, however, it is overkill for
something I maintain in my free time and as an hobby.