chore: merge dnscore into rbmk repository (#76)
The overall purpose for this diff is to simplify the development
environment and reduce chores for me.

My original setup was OK for a professionally maintained project.
However, it is overkill for a hobby project.

This diff is based on
https://github.com/rbmk-project/dnscore/commit/2bb1c489f0f650a9fdfdbf75d60446cfc7650b31.

This diff is based on
https://github.com/rbmk-project/common/commit/443e41e5d38260ff6620be93e98184f41be40402.

This diff includes code written by @roopeshsn as part of
https://github.com/rbmk-project/dnscore/pull/18.